refactor: 불필요한 메서드 삭제 & 반복되는 문자열 로직으로 처리
