+++ src/js/index.js
@@ -1,4 +1,4 @@
 import LottoController from './lottoController.js';
 
 const lottoController = new LottoController();
-lottoController.init();
+lottoController.initLottoGame();
