+++ src/js/models/Lotto.js
